[CHANGE] response version type

+++ src/main/java/usw/suwiki/SuwikiVersion.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class SuwikiVersion {
-    public static final int version = 1;
+    public static final float version = 1;
 }
